requirements.txt
matplotlib
seaborn
gym
mujoco
torchvision
tensorboard
tensorboard-logger
pygame
PyYAML

sudo apt-get install libglfw3
sudo apt-get install libglfw3-dev